backend/requirements_name.txt
fastapi
uvicorn
pydantic-settings
llama-index-core
llama-index-llms-ollama
llama-index-embeddings-huggingface
llama-index-vector-stores-postgres
langchain-community
langchain-text-splitters
psycopg2-binary
requests
